tests: assert shared expectations against both storage services
